feat: add bans

+++ vnd/reports/admin.py
@@ -1,3 +1,5 @@
+import datetime
+
 from django.http import HttpRequest
 from django.shortcuts import redirect
 from django.utils.html import format_html
@@ -6,7 +8,7 @@ from unfold.admin import ModelAdmin
 from unfold.decorators import action
 
 from posts.models import Post
-from reports.models import Report, Reason, ClosedReport
+from reports.models import Report, Reason, ClosedReport, Ban
 
 
 @admin.register(Report)
@@ -14,8 +16,6 @@ class ReportAdmin(ModelAdmin):
     list_display = ["id", "reported_at", "reason", "closed", "content_instance"]
     list_display_links = list_display
     list_filter = ["closed", "reason", "reason__top_priority"]
-    readonly_fields = ["closed", "reason", "reporting_user", "content_instance", "content", "content_model",
-                       "content_id"]
     actions_detail = ["approve"]
 
     def content(self, obj):
@@ -27,11 +27,17 @@ class ReportAdmin(ModelAdmin):
     @action(description="Approve report")
     def approve(self, request: HttpRequest, object_id: int):
         report = Report.objects.get(pk=object_id)
-        report.close(approved=True, admin=request.user)
+        closed_report = report.close(approved=True, admin=request.user)
         report.closed = True
         report.save()
         if report.reason.top_priority:
-            report.content_instance.delete()
+            user = report.reported_user
+            if hasattr(user, 'bans'):
+                user.is_active = False
+            else:
+                ban_end = datetime.date.today() + datetime.timedelta(days=31)
+                report.content_instance.delete()
+                Ban.objects.create(reason=closed_report, user=user, end=ban_end)
         return redirect(f"/admin/reports/")
 
 
@@ -44,3 +50,4 @@ class ClosedReportAdmin(ModelAdmin):
 
 
 admin.site.register(Reason)
+admin.site.register(Ban)
